Work on ledger reconnect
