+++ src/components/ProviderInitializer/helpers/index.ts
@@ -1,0 +1,3 @@
+export * from './getOperaProvider';
+export * from './getExtensionProvider';
+export * from './getHwWalletProvider';
